pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.8"
djongo = "^1.3.4"
django-crispy-forms = "^1.11.2"
pytest = "^6.2.3"
selenium = "^3.141.0"
[tool.poetry.dev-dependencies]
